requirements.txt
blinker
cffi
click
colorama
cryptography
Flask
itsdangerous
Jinja2
MarkupSafe
pycparser
urllib3
Werkzeug
gevent
xmltodict
requests
flask_cors
pymysql
httpx
jieba
requests
pyzbar
Pillow
sqlalchemy
tencentcloud-sdk-python
openai
pyyaml